refactored permissions manager
